improve injection of deps
