+++ app/lib/src/providers/client.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/widgets.dart';
-import 'package:grpc/grpc.dart';
 import 'package:provider/provider.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:generated_io/generated_io.dart';
+import 'package:stashall/src/stores/server.dart';
 
 class ClientProvider extends StatelessWidget {
   final Widget child;
@@ -10,14 +11,19 @@ class ClientProvider extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final channel = Provider.of<ClientChannel>(context);
-    return MultiProvider(
-      providers: [
-        Provider<AuthenticationServiceClient>(
-          builder: (c) => AuthenticationServiceClient(channel),
-        ),
-      ],
-      child: child,
+    final serverStore = Provider.of<ServerStore>(context);
+    return Observer(
+      builder: (c) {
+        final channel = serverStore.channel;
+        return MultiProvider(
+          providers: [
+            Provider<AuthenticationServiceClient>(
+              builder: (c) => AuthenticationServiceClient(channel),
+            ),
+          ],
+          child: child,
+        );
+      },
     );
   }
 }
