+++ app/lib/src/stores/server.dart
@@ -1,3 +1,4 @@
+import 'package:grpc/grpc.dart';
 import 'package:mobx/mobx.dart';
 
 part 'server.g.dart';
@@ -5,9 +6,21 @@ part 'server.g.dart';
 class ServerStore = ServerStoreBase with _$ServerStore;
 
 abstract class ServerStoreBase with Store {
-  @observable
-  String host = '127.0.0.1';
+  static const stashallDefaultHost = '127.0.0.1';
+  static const stashallDefaultPort = 443;
 
   @observable
-  int port = 443;
+  String host = stashallDefaultHost;
+
+  @observable
+  int port = stashallDefaultPort;
+
+  @observable
+  ClientChannel channel = ClientChannel(
+    stashallDefaultHost,
+    port: stashallDefaultPort,
+    options: ChannelOptions(
+      credentials: ChannelCredentials.insecure(),
+    ),
+  );
 }
